Fix bug with personal/config/.gitconfig
The template in core for including personal .gitconfig was
not working correctly for spin environments. This was because
spin makes use of a different dotfiles folder location for
the dotfiles stored under the Shopify/dotfiles repo, to prevent
overwriting the dotfiles pulled from an individual github users
dotfiles repo.

The fix is to include the shopify-dotfiles folder path in the
.gitconfig include block. This pulls in both the personal
and shopify dotfiles as a part of the .gitconfig include.